refactor: directory 생성 및 history style 분리
