+++ src/components/history/_styles/MissionHistoryStyle.ts
@@ -1,0 +1,9 @@
+export const headerStyles = {
+    write: 'bg-[#F1F9F8]',
+    history: 'bg-secondary',
+};
+
+export const headerFontStyles = {
+    write: 'text-secondary',
+    history: 'text-white',
+};
